+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Messages: preserve quoted-message context when replying to stored incoming or outgoing text messages. (thanks @odilorg)
+
 ## 0.12.1 - 2026-07-09
 
 ### Added